moving some functions to core.gradients
